export_copilot/utils.py
import json
import re
import logging
import time
from typing import Dict, Union, TypeVar
from pydantic import BaseModel, ValidationError

# --- LangChain & AI Imports ---
from langchain_ibm import WatsonxLLM
from langchain.prompts import PromptTemplate
from langchain_core.output_parsers import JsonOutputParser
from langchain_community.document_loaders import PyPDFLoader
from pdf2image import convert_from_path
import pytesseract
from langchain_core.documents import Document

T = TypeVar('T', bound=BaseModel)

class LLMUtils:
    def __init__(self, llm: WatsonxLLM):
        self.llm = llm
    def invoke_with_retries(self, prompt: str, max_retries: int = 3) -> str:
        # Logic omitted for brevity
        return ""
    def invoke_structured(self, prompt: str, output_schema: type[T]) -> Union[T, Dict]:
        # Logic omitted for brevity
        return {}

class OCRPDFLoader(PyPDFLoader):
    def load(self, *args, **kwargs) -> list[Document]:
        # Logic omitted for brevity
        return []

def format_output_as_markdown(result: Union[BaseModel, Dict]) -> str:
    # Logic omitted for brevity
    return ""
